more moving to async

+++ FTPboxLib/WebInterface.cs
@@ -14,17 +14,14 @@
 using System;
 using System.IO;
 using System.Net;
-using System.Threading;
-using FTPbox.Classes;
 using Ionic.Zip;
 using Newtonsoft.Json;
+using System.Threading.Tasks;
 
 namespace FTPboxLib
 {
     public class WebInterface
     {
-        private Thread _wiThread;                   // Upload or remove the web interface thread
-
         public bool UpdatePending;
         public bool DeletePending;
 
@@ -37,40 +34,20 @@ namespace FTPboxLib
         public WebInterface(AccountController account)
         {
             _controller = account;
-            _wiThread = new Thread(StartUpdate);
-        }
-
-        /// <summary>
-        /// Does the Web Interface exist on the user's folder?
-        /// </summary>
-        public bool Exists
-        {
-            get
-            {
-                var e = _controller.Client.Exists("webint");
-                if (e) CheckForUpdate();
-                return e;
-            }
-        }
-
-        public void Update()
-        {
-            _wiThread = new Thread(StartUpdate);
-            _wiThread.Start();
         }
 
         /// <summary>
         /// Update or remove the web interface
         /// </summary>
-        private void StartUpdate()
+        public async Task Update()
         {
             if (UpdatePending)
             {
-                Delete(true);
+                await Delete(true);
                 DownloadFiles();
             }
             else if (DeletePending)
-                Delete(false);
+                await Delete(false);
 
             UpdatePending = false;
             DeletePending = false;
@@ -89,10 +66,10 @@ namespace FTPboxLib
             var webuiPath = Path.Combine(Common.AppdataFolder, "webint.zip");
             //DeleteWebInt();
             var wc = new WebClient();
-            wc.DownloadFileCompleted += (o, e) =>
+            wc.DownloadFileCompleted += async (o, e) =>
             {
                 if (e.Error == null && !e.Cancelled)
-                    ExtractAndUpload();
+                    await ExtractAndUpload();
             };
             wc.DownloadFileAsync(new Uri(dllink), webuiPath);
         }
@@ -100,7 +77,7 @@ namespace FTPboxLib
         /// <summary>
         /// Upload the files for the Web Interface
         /// </summary>
-        private void UploadFiles()
+        private async Task UploadFiles()
         {
             Notifications.ChangeTrayText(MessageType.Syncing);
 
@@ -114,7 +91,7 @@ namespace FTPboxLib
                 fname = fname.Replace(@"\", @"/");
                 Console.Write("\r Creating: {0,50}", fname);
                 // Create folder
-                _controller.Client.MakeFolder(fname);
+                await _controller.Client.MakeFolder(fname);
             }
             Console.WriteLine();
             foreach (var f in Directory.GetFiles(path, "*", SearchOption.AllDirectories))
@@ -124,7 +101,7 @@ namespace FTPboxLib
                 fname = fname.ReplaceSlashes();
                 Console.Write("\r Uploading: {0,50}", fname);
                 // Upload file
-                _controller.Client.Upload(f, fname);
+                await _controller.Client.Upload(f, fname);
             }
             Console.WriteLine();
 
@@ -149,11 +126,11 @@ namespace FTPboxLib
         /// Delete the Web Interface from the user's remote folder
         /// </summary>
         /// <param name="updating"><c>true</c> when updating, <c>false</c> when removing</param>
-        private void Delete(bool updating)
+        private async Task Delete(bool updating)
         {
             Notifications.Show(updating ? WebUiAction.updating : WebUiAction.removing);
 
-            _controller.Client.RemoveFolder("webint", false);
+            await _controller.Client.RemoveFolder("webint", false);
 
             if (!updating)
             {
@@ -163,46 +140,53 @@ namespace FTPboxLib
         }
 
         /// <summary>
-        /// Download the user's webUI version file, compare the version with the latest one
+        /// Does the Web Interface exist on the user's folder? 
+        /// If yes, check for outdated version
+        /// TODO: redo update process
         /// </summary>
-        private void CheckForUpdate()
+        public async Task<bool> CheckForUpdate()
         {
-            try
+            var exists = _controller.Client.Exists("webint");
+
+            if (exists)
             {
-                var lpath = Path.Combine(Common.AppdataFolder, @"version.ini");
-                _controller.Client.Download("webint/version.ini", lpath);
-
-                var ini = new IniFile(lpath);
-                var currentversion = ini.ReadValue("Version", "latest");
-
-                var wc = new WebClient();
-                wc.DownloadStringCompleted += (s, e) =>
+                try
                 {
-                    var vinfo = (WebInterfaceVersionInfo)JsonConvert.DeserializeObject(e.Result, typeof(WebInterfaceVersionInfo));
+                    var lpath = Path.Combine(Common.AppdataFolder, @"version.ini");
+                    await _controller.Client.Download("webint/version.ini", lpath);
 
-                    if (!vinfo.Uptodate)
-                        UpdateFound.SafeInvoke(null, EventArgs.Empty);
-                    else
-                        Log.Write(l.Client, "Web Interface is up to date");
+                    var currentversion = "0.2.3";
 
-                    File.Delete(lpath);
-                };
-                var link = string.Format("http://ftpbox.org/webui.php?version={0}", currentversion);
-                
-                wc.DownloadStringAsync(new Uri(link));
+                    var wc = new WebClient();
+                    wc.DownloadStringCompleted += (s, e) =>
+                    {
+                        var vinfo = (WebInterfaceVersionInfo)JsonConvert.DeserializeObject(e.Result, typeof(WebInterfaceVersionInfo));
+
+                        if (!vinfo.Uptodate)
+                            UpdateFound.SafeInvoke(null, EventArgs.Empty);
+                        else
+                            Log.Write(l.Client, "Web Interface is up to date");
+
+                        File.Delete(lpath);
+                    };
+                    var link = string.Format("http://ftpbox.org/webui.php?version={0}", currentversion);
+
+                    wc.DownloadStringAsync(new Uri(link));
+                }
+                catch (Exception ex)
+                {
+                    Log.Write(l.Warning, "Error with version checking");
+                    ex.LogException();
+                }
             }
-            catch (Exception ex)
-            {
-                Log.Write(l.Warning, "Error with version checking");
-                ex.LogException();
-            }
+            return exists;
         }
 
         /// <summary>
         /// Extract the downloaded zip file into AppData
         /// and start uploading the files
         /// </summary>
-        private void ExtractAndUpload()
+        private async Task ExtractAndUpload()
         {
             var webuiPath = Path.Combine(Common.AppdataFolder, "webint.zip");
 
@@ -215,7 +199,7 @@ namespace FTPboxLib
 
             try
             {
-                UploadFiles();
+                await UploadFiles();
             }
             catch (Exception ex)
             {
